InstanceColorization/requirements.txt
absl-py==0.9.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.4b2
cachetools==4.1.0
certifi==2021.10.8
chardet==3.0.4
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.24
decorator==5.1.0
detectron2==0.6+cu102
distlib==0.3.2
dominate==2.4.0
filelock==3.0.12
fonttools==4.28.2
future==0.18.2
fvcore==0.1.5.post20211023
gdown==4.2.0
google-auth==1.14.2
google-auth-oauthlib==0.4.1
grpcio==1.28.1
hydra-core==1.1.1
idna==2.9
imageio==2.11.1
importlib-metadata==1.6.0
iopath==0.1.9
ipython==7.29.0
jedi==0.18.1
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.3.2
Markdown==3.2.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
mock==4.0.2
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.0
omegaconf==2.1.1
opencv-python==4.3.0.38
packaging==21.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
portalocker==1.7.0
prompt-toolkit==3.0.22
protobuf==3.11.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.1
Pygments==2.10.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==5.1
pyzmq==18.1.1
regex==2021.11.10
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.18.3
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.3.1
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchfile==0.1.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.0
urllib3==1.25.8
virtualenv==20.4.7
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
yacs==0.1.8
zipp==3.1.0
